packages/pydantic-ai-absurd/pyproject.toml
[project]
name = "pydantic-ai-absurd"
version = "0.1.0"
description = "Absurd durable execution integration for PydanticAI"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "absurd-sdk>=0.3.0",
    "pydantic-ai-slim[mcp]>=1.18.0",
]

[build-system]
requires = ["uv_build>=0.11.3,<0.12"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.1.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
